+++ sketch-compose/src/main/java/com/github/panpf/sketch/compose/CrossfadePainter.kt
@@ -3,6 +3,8 @@ package com.github.panpf.sketch.compose
 import android.os.SystemClock
 import androidx.compose.runtime.Stable
 import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableFloatStateOf
+import androidx.compose.runtime.mutableIntStateOf
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.geometry.Size
@@ -33,11 +35,11 @@ internal class CrossfadePainter(
     private val preferExactIntrinsicSize: Boolean,
 ) : Painter() {
 
-    private var invalidateTick by mutableStateOf(0)
+    private var invalidateTick by mutableIntStateOf(0)
     private var startTimeMillis = -1L
     private var isDone = false
 
-    private var maxAlpha: Float by mutableStateOf(DefaultAlpha)
+    private var maxAlpha: Float by mutableFloatStateOf(DefaultAlpha)
     private var colorFilter: ColorFilter? by mutableStateOf(null)
 
     override val intrinsicSize get() = computeIntrinsicSize()